JSON PLACE + SAVE URL + FILE CHECK KEY

+++ post_send.py
@@ -1,7 +1,7 @@
 import requests
 from tkinter import filedialog
 import tkinter as tk
-from tkinter import Menu
+from tkinter.ttk import Combobox
 import json
 import pyperclip
 
@@ -53,14 +53,18 @@ class Programm():
 		self.button_data_delete = tk.Button(root, text = "Удалить", command =self.data_delete)
 		self.button_send_data_POST = tk.Button(root, text = "Отправить POST", command =self.send_POST)
 		self.button_send_data_GET = tk.Button(root, text = "Отправить GET", command =self.send_GET)
-
+		self.button_urls_save = tk.Button(root, text = "Добавить ссылку", command =self.urls_save)
 		self.button_test = tk.Button(root, text = "ТестКнопка", command =self.test)
 		
 		self.list_box = tk.Listbox()
 		self.list_box_2 = tk.Listbox()
 
+		self.button_place_data = tk.Button(root, text = "Вставить данные", command = lambda: PlaceText(self.list_box_2))
+
+		self.combox = Combobox(root)
+
 		self.L0.grid(row = 0, column = 2)
-		self.L1.grid(row = 3, column = 2, rowspan = 1)
+		self.L1.grid(row = 3, column = 2)
 		self.L2.grid(row = 1, column = 2)
 		
 		self.url.grid(row = 1, column = 1, padx = 20)
@@ -71,6 +75,10 @@ class Programm():
 
 		self.button_send_data_POST.grid(row = 2, column = 1)
 		self.button_send_data_GET.grid(row = 3, column = 1)
+		self.button_urls_save.grid(row = 4, column = 1)
+		self.combox.grid(row = 5, column = 1, rowspan = 3)
+
+		self.button_place_data.grid(row = 6, column = 1)
 
 		self.button_data_edit.grid(row = 2, column = 1, columnspan = 2)
 		self.button_data_add.grid(row = 2, column = 2, columnspan = 1)
@@ -83,8 +91,11 @@ class Programm():
 
 		self.dirs.insert(0,'Выберите файл')
 		self.url.insert(0,'http://httpbin.org/post') # https://truecoding.ru/guild/pvp2.php
+
 		
 		self.list_box_2.bind("<<ListboxSelect>>", self.focusbox)
+		self.combox.bind("<<ComboboxSelected>>", self.focuscombox)
+
 		root.protocol("WM_DELETE_WINDOW", self.save_session)
 
 		json_data_str = open("session.json")
@@ -122,8 +133,19 @@ class Programm():
 						self.list_box_2.delete(0, tk.END)
 						for data_list in json_data[data_json][::-1]:
 							self.list_box_2.insert(0, data_list)
+					elif data_json == "urls":
+						self.combox['values'] = json_data[data_json]
 		except json.decoder.JSONDecodeError:
 			pass
+
+	def urls_save(self):
+		url = self.url.get()
+		urls = []
+		urls.append(url)
+		for value in self.combox["values"]:
+			urls.append(value)
+		self.combox["values"] = urls
+
 	def focusbox(self,get):
 		cursor = self.list_box_2.curselection()
 		data_all = self.list_box_2.get(cursor, cursor)
@@ -147,6 +169,10 @@ class Programm():
 				self.data.insert(0, data_key)
 				self.data_value.insert(0, data_value)
 
+	def focuscombox(self,get):
+		cursor = self.combox.get()
+		self.url.delete(0, tk.END)
+		self.url.insert(0, cursor)
 
 	def file_dir(self):
 		file = filedialog.askopenfilename()
@@ -158,14 +184,19 @@ class Programm():
 		self.name.insert(0, f"{file.split('/')[-1]}")
 
 
-	def Check_key(self):
+	def Check_key(self, data):
 
-		data_key1 = self.data.get()
+		data_key1 = data
+
+		print(data_key1)
 
 		data = self.list_box_2.get(0, tk.END)
 
 		for data_key_value in data:
 			data_key2 = data_key_value.split(' : ')[0]
+			if data_key2 == 'file':
+				data_key = data_key_value.split(' : ')[1]
+				data_key2 = data_key.split(' --> ')[0]
 			if data_key1 == data_key2:
 				self.L1['text'] = "Ошибка Такие данные уже есть в таблице пожалуйста введите другое название"
 				return 1
@@ -174,29 +205,39 @@ class Programm():
 		file = self.dirs.get()
 		name = self.name.get()
 
-		self.dirs.delete(0, tk.END)
+		if self.Check_key(name) == 1:
+			return
+
 		self.name.delete(0, tk.END)
+		self.dirs.delete(0, tk.END)
 
 		if name != "":
-			self.list_box_2.insert(0, f'file : {name} --> {file}')
+			if self.Check_key(name) != 1:
+				self.list_box_2.insert(0, f'file : {name} --> {file}')
+			else:
+				return
 		else:
 			if file == "" or file == "Выберите файл":
 				self.L1['text'] = "Ошибка Выберите файл пожалуйста"
 				return
 			else:
-				self.list_box_2.insert(0, f'file : {file.split("/")[-1]} --> {file}')
+				file = file.split('/')[-1]
+				if self.Check_key(file) == 1:
+					return
+				else: 
+					self.list_box_2.insert(0, f'file : {file.split("/")[-1]} --> {file}')
 
 	def data_add(self):
 		data = self.data.get()
 		data_value = self.data_value.get()
 
-		self.data.delete(0, tk.END)
-		self.data_value.delete(0, tk.END)
-
-		check = self.Check_key()
+		check = self.Check_key(data)
 
 		if check == 1:
 			return
+
+		self.data_value.delete(0, tk.END)
+		self.data.delete(0, tk.END)
 
 		if data == "" and data_value == "":
 			self.L1['text'] = "Ошибка Введите данные"
@@ -213,7 +254,8 @@ class Programm():
 	def data_edit(self):
 		index = self.list_box_2.curselection()
 		data = self.list_box_2.get(index, index)
-		key = data.split(' : ')[0]
+		print(data)
+		key = data[0].split(' : ')[0]
 
 		if key == "file":
 			dirs = self.dirs.get()
@@ -371,7 +413,10 @@ class Programm():
 		data_value = self.data_value.get()
 		data_listbox1 = self.list_box.get(0, tk.END)
 		data_listbox2 = self.list_box_2.get(0, tk.END)
-
+		urls = []
+		for value in self.combox["values"]:
+			urls.append(value)
+		
 		if data == "":
 			data = None
 
@@ -389,6 +434,7 @@ class Programm():
 
 		json_data = {}
 		json_data["url"] = url
+		json_data["urls"] = urls
 		json_data["dirs"] = dirs
 		json_data["name"] = name
 		json_data["data"] = data
@@ -403,6 +449,41 @@ class Programm():
 		file.close()
 
 		root.destroy()
+class PlaceText:
+	def __init__(self, listbox2):
+		self.rooter = tk.Toplevel(root)
+		self.rooter.title('Place')
+		self.listbox2 = listbox2
+
+		self.TextEntry = tk.StringVar()
+
+		self.text = tk.Text(self.rooter, wrap = tk.WORD)
+		self.btn = tk.Button(self.rooter, text = "Place", command = self.place)
+		self.btn_text = tk.Button(self.rooter, text = "Вставить из буфера обмена", command = self.place_text)
+
+		self.text.grid(row = 1, column = 1)
+		self.btn.grid(row = 2, column = 1)
+		self.btn_text.grid(row = 2, column = 2)
+
+	def place_text(self):
+		q = pyperclip.paste()
+		self.text.insert(1.0, q)
+
+	def place(self):
+		text_place = self.text.get(1.0, tk.END)
+		json_data = json.loads(text_place)
+
+		self.listbox2.delete(0, tk.END)
+
+		for data,value in json_data.items():
+			self.listbox2.insert(0, f"{data} : {value}")
+			
+		def __missing__(self, key):
+			pass
+
+		return self.rooter.destroy()
+
+
 s = Programm(root)
 
 root.mainloop()
